Update PS1_Systray_Tool.ps1
I added the update found in the following comment from an unknown user.
--------------------
This is an amazing script. Thank you for the walk through.
I have noticed that when you close the Window and exit you haven't ended the appcontext thread. My powershell was stating I hadn't shut down cleanly.

The small change I have made was to the Menu_Exit script and Menu_Restart_Tool as below:
$Main_Tool_Icon.Visible = $false
$window.Close()
$appContext.ExitThread()
Stop-Process $pid